+++ src/test/java/ch/bitagent/bitcoin/lib/helper/PropertiesTest.java
@@ -7,16 +7,6 @@ import static org.junit.jupiter.api.Assertions.*;
 class PropertiesTest {
 
     @Test
-    void getBlockstreamMainnetUrl() {
-        assertEquals("https://blockstream.info/api", Properties.getBlockstreamMainnetUrl());
-    }
-
-    @Test
-    void getBlockstreamTestnetUrl() {
-        assertEquals("https://blockstream.info/testnet/api", Properties.getBlockstreamTestnetUrl());
-    }
-
-    @Test
     void getWallets() {
         assertEquals(2, Properties.getWallets(Properties.WALLET_FILENAME).size());
         assertThrowsExactly(IllegalStateException.class, () -> Properties.getWallets("bla"));